# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  backend_internal_model.ChangePasswordReq:
    properties:
      newPassword:
        maxLength: 64
        minLength: 6
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  backend_internal_model.CreateTagReq:
    properties:
      name:
        maxLength: 64
        type: string
    required:
    - name
    type: object
...
  backend_internal_model.Tag:
    properties:
      createdAt:
        type: string
      deletedAt:
        format: date-time
        type: string
      id:
        type: integer
      name:
        type: string
      postCount:
        type: integer
      updatedAt:
        type: string
    type: object
...
  backend_internal_model.UpdateProfileReq:
    properties:
      about:
        maxLength: 500
        type: string
      address:
        maxLength: 256
        type: string
      city:
        maxLength: 64
        type: string
      country:
        maxLength: 64
        type: string
      displayName:
        maxLength: 64
        type: string
      isPublic:
        type: boolean
      phoneNumber:
        maxLength: 32
        type: string
      photoURL:
        maxLength: 512
        type: string
      state:
        maxLength: 64
        type: string
      zipCode:
        maxLength: 16
        type: string
    type: object
  backend_internal_model.UpdateTagReq:
    properties:
      name:
        maxLength: 64
        type: string
    required:
    - name
    type: object
...
  backend_internal_model.UserResponse:
    properties:
      about:
        type: string
      address:
        type: string
      city:
        type: string
      country:
        type: string
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: integer
      isPublic:
        type: boolean
      phoneNumber:
        type: string
      photoURL:
        type: string
      role:
        type: string
      state:
        type: string
      status:
        type: string
      username:
        type: string
      zipCode:
        type: string
    type: object
  backend_internal_pkg.PageResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/backend_internal_pkg.PageResult'
      msg:
        example: success
        type: string
    type: object
  backend_internal_pkg.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  backend_internal_pkg.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 基于 Gin + GORM + SQLite 的个人博客后端接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 个人博客 API
  version: "1.0"
paths:
...
  /admin/tags:
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.CreateTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.Tag'
              type: object
      security:
      - BearerAuth: []
      summary: 创建标签
      tags:
      - tags
  /admin/tags/{id}:
    delete:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 删除标签
      tags:
      - tags
    put:
      consumes:
      - application/json
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.UpdateTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.Tag'
              type: object
      security:
      - BearerAuth: []
      summary: 更新标签
      tags:
      - tags
  /admin/users:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - description: 角色筛选
        in: query
        name: role
        type: string
      - description: 状态筛选
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.PageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 用户列表（管理员）
      tags:
      - admin
  /admin/users/{id}:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 获取指定用户（管理员）
      tags:
      - admin
...
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.UpdateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 更新当前用户信息
      tags:
      - users
  /me/password:
    put:
      consumes:
      - application/json
      description: 修改成功后需重新登录
      parameters:
      - description: 密码信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_model.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - users
...
  /tags:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_pkg.PageResponse'
      summary: 标签列表（公开）
      tags:
      - tags
  /tags/{id}:
    get:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_model.Tag'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_pkg.Response'
      summary: 标签详情（公开）
      tags:
      - tags
...
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"